move serialize_output to cairo-io-serde crate
