+++ cairo-io-serde/src/lib.rs
@@ -1,2 +1,15 @@
-pub mod schema;
+use cairo_vm::Felt252;
+
 pub mod cairo_input;
+pub mod cairo_output;
+pub mod schema;
+
+#[allow(dead_code)]
+#[derive(Debug, PartialEq, Clone)]
+pub enum FuncArg {
+    Array(Vec<Felt252>),
+    Single(Felt252),
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct FuncArgs(pub Vec<FuncArg>);
